// src/api/types/user.api.types.ts
export type User = {
	name: string;
	phone: string;
	email: string;
	admissionDate: string;
	position: string;
	division: string;
	additionalInfo: string;
};
export type UserRequest = {
	name: string;
	phone: string;
	email: string;
	address: string;
	position_name: string;
	department: string;
	hire_date: string;
};